2005-04-17  Zoltan Varga  <vargaz@freemail.hu>
	* aot.c (mono_compile_assembly): Fix warning.

	* mini-exceptions.c (ves_icall_get_frame_info): Fix bug introduced
	by the _MSC_VER stuff.

svn path=/trunk/mono/; revision=43141


Commit migrated from https://github.com/mono/mono/commit/a5b6cf8dbb0c6b77f9511701d66f8ea3178b5dc9

+++ src/mono/mono/mini/aot.c
@@ -2211,7 +2211,7 @@ mono_compile_assembly (MonoAssembly *ass, guint32 opts, const char *aot_options)
 	if (acfg->got_offset > 0)
 		fprintf (tmpfp, ".skip %d\n", (int)(acfg->got_offset * sizeof (gpointer)));
 
-	printf ("GOT SIZE: %d\n", acfg->got_offset * sizeof (gpointer));
+	printf ("GOT SIZE: %d\n", (int)(acfg->got_offset * sizeof (gpointer)));
 
 	symbol = g_strdup_printf ("got_addr");
 	emit_section_change (tmpfp, ".data", 1);
